shots measurements now method of MPS
